+++ Cargo.toml
@@ -22,7 +22,6 @@ rustdoc-args = ["--cfg", "docsrs"]
 [dependencies]
 async-trait = "0.1.57"
 axum = { version = "0.6.20", default-features = false }
-futures = "0.3.28"
 http = "0.2.9"
 serde = "1"
 thiserror = "1.0.49"
@@ -39,7 +38,7 @@ hyper = "0.14.27"
 tokio = { version = "1.32.0", features = ["full"] }
 tower = "0.4.13"
 tokio-test = "0.4.3"
-sqlx = { oversion = "0.7.2", features = [
+sqlx = { version = "0.7.2", features = [
     "sqlite",
     "time",
     "uuid",
