+++ src/lib.rs
@@ -7,10 +7,13 @@
 #![forbid(unsafe_code)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+mod auth;
 mod extract;
+mod middleware;
 mod service;
 mod user_store;
 
-pub use extract::{require_authn, require_authz};
-pub use service::{Auth, LoginManager, LoginManagerLayer};
+pub use auth::Auth;
+pub use middleware::{require_authn, require_authz};
+pub use service::{LoginManager, LoginManagerLayer};
 pub use user_store::UserStore;
